move authentification helpers into a non-api-route directory
